ai/ai-vision/video-stream-analysis/requirements.txt
streamlit
pillow
numpy
opencv-python
pandas
altair
oci
vision-service-python-client